Foundation-model-ci-statistics/requirements.txt
gitpython
tabulate
dotenv
pandas
seaborn
matplotlib